use BUCKET_NAME environment variable

+++ Rakefile
@@ -1,7 +1,11 @@
 task :default do
   raise "REMOTE_HOST is not set" unless host = ENV['REMOTE_HOST']
+  raise "BUCKET_NAME is not set" unless ENV['BUCKET_NAME']
+  raise ".s3cfg doesn't exist" unless File.exist?("#{ENV['HOME']}/.s3cfg")
+
   system! %[rsync -az --delete --exclude .git -e ssh . #{host}:soloist]
   system! %[rsync -az -e ssh ~/.s3cfg #{host}:]
+  system! %[ssh -t #{host} "echo $BUCKET_NAME >~/bucket_name"]
   system! %[ssh -t #{host} "cd soloist && sh -ex bootstrap.sh"]
 end
